+++ src/front/js/store/flux.js
@@ -1,6 +1,8 @@
 const getState = ({ getStore, getActions, setStore }) => {
 	return {
 		store: {
+			token: null,
+      		isLoggedIn: false,
 			message: null,
 			demo: [
 				{
@@ -34,6 +36,7 @@ const getState = ({ getStore, getActions, setStore }) => {
 					  "https://3001-violetapint-choremanage-wth2flooiwp.ws-eu52.gitpod.io/api/signup",
 					  opts
 					);
+					
 				   
 					if (resp.status !== 200) {
 					  alert("error before initial 200 request");
@@ -56,18 +59,18 @@ const getState = ({ getStore, getActions, setStore }) => {
 			// 	getActions().changeColor(0, "green");
 			// },
 
-			getMessage: async () => {
-				try{
-					// fetching data from the backend
-					const resp = await fetch(process.env.BACKEND_URL + "/api/hello")
-					const data = await resp.json()
-					setStore({ message: data.message })
-					// don't forget to return something, that is how the async resolves
-					return data;
-				}catch(error){
-					console.log("Error loading message from backend", error)
-				}
-			},
+			// getMessage: async () => {
+			// 	try{
+			// 		// fetching data from the backend
+			// 		const resp = await fetch(process.env.BACKEND_URL + "/api/hello")
+			// 		const data = await resp.json()
+			// 		setStore({ message: data.message })
+			// 		// don't forget to return something, that is how the async resolves
+			// 		return data;
+			// 	}catch(error){
+			// 		console.log("Error loading message from backend", error)
+			// 	}
+			// },
 			changeColor: (index, color) => {
 				//get the store
 				const store = getStore();
